feat(native): Automatic rollout to highest priority kit
